feat: Telegram Mini App (WebApp)
- FastAPI backend (webapp/) with 14 REST API endpoints
  - Auth: POST /api/auth (Telegram initData HMAC verification)
  - User: GET /api/user/me, PATCH /api/user/language, GET /api/user/orders
  - Catalog: cities, locations (with stock summary), products, categories
  - Cart: full CRUD with quantity management
  - Orders: create order + admin bot notifications

- React + Vite + TypeScript frontend (frontend/)
  - Dark branded design (purple/cyan theme)
  - Pages: Home, Cities, Locations, Products, ProductDetail, Cart, Checkout, Profile, OrderSuccess
  - Checkout FSM: delivery type -> contact -> date/time -> payment
  - Zustand state: cart + user stores
  - Bottom navigation with cart badge
  - Telegram WebApp SDK integration (expand, haptic, MainButton)

- Bot updates: /start now shows WebApp launch button
- Deploy config: railway.toml, Procfile, frontend/vercel.json
- Frontend builds clean: 271kb JS + 4.7kb CSS

+++ db/models/cart.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from sqlalchemy import BigInteger, ForeignKey
+from sqlalchemy import ForeignKey, Integer
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from db.session import Base
@@ -9,13 +9,13 @@ from db.session import Base
 class Cart(Base):
     __tablename__ = "carts"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     user_id: Mapped[int] = mapped_column(
-        BigInteger, ForeignKey("users.id", ondelete="CASCADE"), nullable=False, unique=True, index=True
+        Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False, unique=True, index=True
     )
     # Which location the user is browsing (None = InPost mode)
     location_id: Mapped[int | None] = mapped_column(
-        BigInteger, ForeignKey("locations.id", ondelete="SET NULL"), nullable=True
+        Integer, ForeignKey("locations.id", ondelete="SET NULL"), nullable=True
     )
 
     user: Mapped["User"] = relationship("User", back_populates="cart")
