+++ db/models/cart_item.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from sqlalchemy import BigInteger, ForeignKey, Integer, UniqueConstraint
+from sqlalchemy import ForeignKey, Integer, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from db.session import Base
@@ -10,12 +10,12 @@ class CartItem(Base):
     __tablename__ = "cart_items"
     __table_args__ = (UniqueConstraint("cart_id", "variant_id", name="uq_cart_variant"),)
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     cart_id: Mapped[int] = mapped_column(
-        BigInteger, ForeignKey("carts.id", ondelete="CASCADE"), nullable=False, index=True
+        Integer, ForeignKey("carts.id", ondelete="CASCADE"), nullable=False, index=True
     )
     variant_id: Mapped[int] = mapped_column(
-        BigInteger, ForeignKey("product_variants.id", ondelete="CASCADE"), nullable=False
+        Integer, ForeignKey("product_variants.id", ondelete="CASCADE"), nullable=False
     )
     quantity: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
