+++ db/models/location_stock.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from sqlalchemy import BigInteger, DateTime, ForeignKey, Integer, UniqueConstraint
+from sqlalchemy import DateTime, ForeignKey, Integer, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from db.session import Base
@@ -14,12 +14,12 @@ class LocationStock(Base):
     __tablename__ = "location_stock"
     __table_args__ = (UniqueConstraint("location_id", "variant_id", name="uq_location_variant"),)
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     location_id: Mapped[int] = mapped_column(
-        BigInteger, ForeignKey("locations.id", ondelete="CASCADE"), nullable=False, index=True
+        Integer, ForeignKey("locations.id", ondelete="CASCADE"), nullable=False, index=True
     )
     variant_id: Mapped[int] = mapped_column(
-        BigInteger, ForeignKey("product_variants.id", ondelete="CASCADE"), nullable=False, index=True
+        Integer, ForeignKey("product_variants.id", ondelete="CASCADE"), nullable=False, index=True
     )
     quantity: Mapped[int] = mapped_column(Integer, nullable=False, default=0, server_default="0")
     last_sold_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
