+++ db/models/order_item.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from decimal import Decimal
 
-from sqlalchemy import BigInteger, ForeignKey, Integer, Numeric
+from sqlalchemy import ForeignKey, Integer, Numeric
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from db.session import Base
@@ -11,12 +11,12 @@ from db.session import Base
 class OrderItem(Base):
     __tablename__ = "order_items"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     order_id: Mapped[int] = mapped_column(
-        BigInteger, ForeignKey("orders.id", ondelete="CASCADE"), nullable=False, index=True
+        Integer, ForeignKey("orders.id", ondelete="CASCADE"), nullable=False, index=True
     )
     variant_id: Mapped[int | None] = mapped_column(
-        BigInteger, ForeignKey("product_variants.id", ondelete="SET NULL"), nullable=True
+        Integer, ForeignKey("product_variants.id", ondelete="SET NULL"), nullable=True
     )
     quantity: Mapped[int] = mapped_column(Integer, nullable=False)
     price_at_order: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
